Move the definition of equals and _struct_equals from Object to its C++-level superclass Reference (which is the common C++ supertype for all heap-allocated X10 values (objects, closures, boxed structs).
A more involved commit later today will generate an overriding
definition of _struct_equals for each closure literal that will
actually implement the desired equality semantics on functions.  
For now,_struct_equals is inherited from Reference, and is defined to
be pointer equality.  Making this interim change so that any front-end
changes needed to allow closures to be used in equals/== can be made
without waiting for the codegen changes for a proper _struct_equals for
closure literals,