A less ambituous version of r12782 that only attempts to make the MPI-based variant of x10rt into a shared library.
This is sufficient to enable the development of Java bindings for
X10RT to proceed without requiring that PGAS always be built in 
a way that supports shared libraries and/or libtool.  We will eventually
need to revisit this and change how we build/package PGAS for X10
so that shared library versions of PGAS are built & included in
the PGAS binary tarballs that are input into the X10RT build.