Take two at codegen for @StackAllocate.  The previous attempt was not correct because the C++ compiler is smart enough to re-use the stack space for variable declated inside of a statement expression.  So, we instead support exactly the usage pattern of:
 @StackAllocate val v = @StackAllocate new T(...);

The @StackAllocate annotation on the variable declaration causes the
compiler to declare a temporary variable of type T (not ref<T>)
to get the necessary stack space.  The name of this variable is 
stored in the context and when the new is encountered, instead of 
calling _make the compiler instead calls _constructor on the tmp variable.
After the init_expr, the address of the tmp variable is taken and stored
in a local variable of ref type.

For example:

    @StackAllocate val f1 = @StackAllocate new Foo(10,20); 

yields the C++ code:

    Foo _StackAllocate_f1; _StackAllocate_f1._constructor(((x10_int)10), ((x10_int)20));
    x10aux::ref<Foo> f1 = (&_StackAllocate_f1);