XTENLANG-930: Rail.toString() should print rail contents
The C++ and Java native implementations of Rail had gotten out
of synch.  The C++ version has printed something useful for a while,
the Java version was just delegating to toString() on the backing
Java array object, which does not print its elements.  

So, this commit changes the Java implementation to match the C++ one.
We could consider a slightly smarter implementation that only prints out
the first few/last few elements if the array Rail is above a size threshold
(for example print first 10, then "...n-20 elements..." then  last 10).