This is an interesting augmentation of the type checker.
Consider the expression e as T. The type checker now infers the type of this expression to be T{self==v} if the selfvar of the type of e is bound to v, and the cast is either an unchecked cast (type of e is a subtype of T) or a checked cast (with no user-defined conversions).

See XTENLANG-1053 for a program that type-checks because of this additional rule.