Ensure that casts are generated for weak closure calls.
Added -VERBOSE_CALLS compiler option. This prints out a warning for each weak call, together with the type the expression was cast to.

Added -STRONG_CALLS compiler option. This causes the compiler to give an error when it encounters a weak call, instead of generating a cast.

The default (no -VERBOSE_CALLS, no -STRONG_CALLS) is to print out a count of weak calls.