straggling files.
had to insert a few casts that shd not be needed (e.g. for dits | r).

Added a constraint on the constructor for RectRegion that the two ValRails are of the same size. 

For those cases in which the size is an expression, simply create a local variable to hold the expression, and use that variable when making the ValRails. 

Removed some constraints that seem to be incorrect (some (rank) clauses in return types of methods on Dist).