Fix AtEach so it does not generate a new place each time it is visited. This helps reduce the situations in which the AST synthesizer operating on a constraint comes across a synthesized place term which is not the same as the current placeTerm stored in the context. (To follow this it helps to know that the type T! for a local variable is implemented by replacing here with the current place term. So without this fix in a second traversal yo would have a place term different from the one that was recorded in the type for the local variable.)
Synthesis of ASTs for type constraints which refer to generated place variables different from the placeTerm in the current context will still throw an errror. Dynamic call conversion will fail for such a situation. This is a limitation. The compiler now generates an error saying there is a place type problem.