Restructure top-level loop to accomplish two things:
  (1) only create the DistArrays once, instead of 
      once per iteration.  This is important because
      when the DistArray is GCed, we don't automatically
      reclaim the backing data for the array at each place.
      We should add that facility to DistArray, but it would
      take a fair amount of plumbing to get in place.
  (2) Only serialize the needed segment of the long String to
      each place instead of serializing the entire String
      and then selecting the needed substring at the place.
      Doesn't matter too much for smallish inputs, but will matter
      as the input sets get larger (O(N) vs. O(N^2) network traffic).

As a secondary benefit, this changes the code to only do the input
phase once instead of once per every iteration, but this is not
actually that significant, since input was a very small fraction of 
execution time.  However, this was the most natural way to structure 
the code to get the O(N) serialization traffic, which I think will
become important as the input scales.