Adapt KMeansDataGenerator to generate data files in the "legacy" format expected by KMeansSPMD.
I think we should consider changing the format isn't quite 
as self-describing as it could be and also has an extra byte-reversal
vs. the way one would write such a binary data file in Java.

Would be nice if we could use the same data files for both X10 and Java
versions of KMeans and minimize the byte swapping that happens.