Workaround problem discovered by Vijay where the C++ backend failed to initialize blank instance fields of X10 struct types that don't correspond to Java built-in primitives.  The workaround is to explicitly call memset in the _make method of classes to zero the storage allocated by x10aux::alloc.
Intentionally didn't insert the memset inside of x10aux::alloc because
in most cases we don't need to zero initialize the storage, so we don't
want to pay for a redundant initialization.

Opened XTENLANG-1407 to remind us to revisit this code once we
have specified the language such that we can get a definite
assignment (or a definite default value) for all fields in the 
object. The combination of that and proto should allow us to
enable the memset as unnecessary.