XTENLANG-1387: Random.random() not actually random.
Patch and comment from Josh:
The problem lies in the implementation of the "RAND" PRNG used 
to seed the initial buffer. This should be a linear congruential generator:

 mt(n) = (69069 * mt(n-1) + 1) mod 2^32 

however the constant factor +1 was omitted. With the addition of the +1, 
the even bias and cyclic behavior disappears.
(See e.g. Vattulainen et al. (1995). "A comparative study of some pseudorandom number generators".)