WIP on XTENLANG-1371
Fix comment on Rail.make (actual subject of issue).

Put in native-level plumbing to allow Rail initialization 
with a specified default value (in addition to an init closure).
API is currently commented out because it causes some
interesting overloading problems (for Rails where T is actually
a closure).  I think those are resolvable, but want to look at it in
a separate small commit.

Also refactored the code to avoid overloading the make
methods with and without alignment, because as soon as we have
a make that takes a default value, making a Rail of int 
becomes dangerously ambiguous (make[int](10,8)...rail of size 10
aligned to 8 or rail of size 10 initialized to 8???).