Fix distribution bug -- there was a confusion in the code between the incoming lifeline edges and outgoing lifeline edges.
The thief knows when it is requesting help from a lifeline. So give trySteal a parameter which records whether the theft request is a lifeline request or not. Now, when a lifeline request is received, push it onto the thieves stack. There is no need to check it against outgoing lifelines. When there is enough loot, send a portion to a recorded thief.