Found source of crashes.
Almost certainly caused by out of memory problem, that was why it was not manifested by an exception at the X10 level.

We want to maintain the invariant that a place P has at most one lifeline request outstanding from those thieves which have P as their lifeline buddy. Otherwise a thief may spam a lifeline buddy with many outstanding requests (potentially causing it to run out of memory to record all these requests).

How can these arise? The thief runs out of work, so tells each lifeline buddy to send it work. It gets work from some buddy. Now it runs out of work again. It again tells every buddy to send it work. Now the invariant is violated.

We guard against this violation by introducing lifelineActivated -- a boolean for each target place. This is set when a request is sent, and reset when loot is received from that buddy.