Fix at least one cause of itable assertion failures that may arise fairly commonly when passing closure literals as arguments to functions.
If the types don't match exactly, the C++ backend does a 
class_cast_unchecked to do the type conversion.  Although it 
is true we don't have to actually check dynamically, it is not
the case in our current code-generation strategy that an
unchecked cast can be a complete no-op.  In particular, we
still have to force the getRTT<T>() class to happen, because that
is the only mechanism we currently have to ensure that the RTT 
is initialized before we get into an itable dispatch sequence.
With the way we currently set up the itables, by the time we get
to the itable dispatch, we no longer have the static information
required to initialize the RTT.

A longer term fix may be to add an extra entry to each itable
that is the address of the getRTT<T> function that needs to be 
called if the itable is uninitialized.  This would allow us to
restore class_cast_unchecked to being a complete no-op with
a relatively minor extension to the itable codegen scheme.
Will open an issue pointing to this commit to consider that
additional work item.