convert RectLayout from a class to a struct.
Closes 1/3 of the performance gap between the Rail and Array based
versions of SSCA1.  Still 20% left, which I think will require the
val fields of a struct to be declared const at the C++ level before it
can be closed. g++ is not optimizing away redundant index calculations
on array accesses because the stride is being loaded from a non-const field
of RectLayout.