Fix closure serialization broken in r15527.
It is critically important that buf.record_reference(obj) is
called BEFORE the body of the closure is deserialized.  If this
is not done and a value captured by the closure (transitively)
refers to an object that has already been deserailized, then the
relative encoding of the previous object will be off by one.

This in turn will result in improper (and not type safe) assignment
of points, which with moderate probability will cause memory 
corruption, crashes, or other bizarre behavior.

I believe this fix will clear up a number of test case regressions
in the C++ backend that happened in the week or so period in mid-August
when regression test machines were done.  It fixed HeatTransfer_v4
and HeatTransfer_v5 (tested and verified), so there is reason to 
believe it will fix other things as well.