This commit refines #16678.
It replaces the @TempNoInline annotation with four annotations allow the system to build while implicitly inlining static or final methods containing up to 0, 1, 2, or 3 calls.  No annotations are required to allow builds with explicit inlining.

It looks as if these annotations mask the effects of two different bugs.  One has to do with improper instantiation of methods with generic parameters.  (To test the effect of removing an annotation, remove the "extends NoInline" clause from its declaration.)