Refactored some type-checking code from AST nodes that was being called from multiple AST nodes into Checker.java.
Also added support for Region.* (multiply) in the type checker. The typechecker now correctly infers the rank of the product region is the sum of the ranks of the input regions. It also propagates rect and zeroBased properties. 

This is very helpful in permitting programs like the following

val e = 0..10;
val r = e*e; // now inferred to be of rank 2. 

It becomes necessary to support this so that the more verbose new Array[Region(1){rect}][e,e] does not need to be used.