Flipped the way array literals are typed, so that the common case is supported better.
The type of an array literal [a, b, ...] is the least upper bound of the *baseTypes* of a, b etc. (used to be the least upper bound of the types of a, b etc)

If you want to preserve the constraints on a,b etc, then use the explicitly typed array literal, new Array[T][a,b,c...]. Here you have to specify the type you want. Each of a, b, c will be checked to have this type. 

Thus now:
[0] // type is Array[Int](1) .. should become Array[Int]{rail}
new Array[int{self==0}][0] // type is Array[int{self==0}](1)