Revert 17046 due to build breakage.
In particular, it causes methods on unconstrained generic types
to be resolved to Object instead of Any.  This in turn results 
in the wrong @Native annotations being generated, which in turn
cases the C++ backend to fail building XRX because an ArrayList
is instantiated on int and that attempts to call Object's equals
method instead of Any's equals method on an int.

We have to find the methods of Any when we have something of
type Any...