XTENLANG-1898 : toString, etc. need to work on captured_lvals
By defining operators on captured_ref_lval and captured_struct_lval
almost all operations "just work."  However, in places where we are using
templates to do static dispatching we need to extend the templates to
handle the new cases introduced by these two new top-level C++ static types.

Tried a number of variations on how best to do this, and believe this is 
the cleanest and most robust way of doing it.