More munging with linker command line options
x10rt tests now work on AIX with both xlC and g++
for standalone, sockets, and lapi transports.

Key AIX-specific issue is that dynamic linking is 
much less permissive than on linux.  One needs to
give the proper absolute paths and lib dependencies
both when making the final executable and when building
.so that depend on other .so (for example libx10rt_pgas_lapi.so).