Bug fix in port of application to 2.1 object model.
The PlaceLocalHandles for local_new_clusters and local_cluster_counts
have to be accessed in the place where they live.  Accessing the 
PlaceLocalHandles in Place 0 where they are going to be accumulated
results in accumulating Place 0's local results N times instead of 
accumulating the results from all the places!