Addressed two long-standing constant propagation problems.
First, declarations of variables with constant initializers were being eliminated, but uses of the variables were not being replaced.  (This is a bug fix.)  

Second, constant propagation in conditionals can make subsequent code unreachable.  This version introduces a hack that makes such code appear to a Java compiler as if it might be reachable.  (The long-term fix is to do dead code elimination after constant propagation.)