fix for broken trunk
Inlining is now off by default.  Three kinds of inlining are supported: method, closure, and (compile-time) constant.  Each has a separate configuration variable that must be true for the optimization to be performed.  Currently, no inlining will be performed unless the EXPRERIMENTAL configuration variable is also set.  The optimizer has a public static final boolean constant (INLINING) that is true if some form of inlining is being performed.

CLOSURE_INLINING must be true if closures are to be inlined.  But, this configuration variable really means that StmtExpr's can be handled by the backend.  It should be renamed (and set to false for the Java backend).