Add pragma for the pure attribute.
Turns out this actually isn't useful for helping the C++ 
compiler optimize itable dispatches because of thew way the 
code is structured in an inline fast path and an out-of-line
slow path.  However, since I looked up what the name of the
atribute is, will add it in case it is useful in the future.