XTENLANG-2100 : Bug in repeated serialization of struct whose static type is an interface
The smallest fix for this particular bug was to add a call to record_reference
in IBox<T>'s deserialize_body method.

However also made significant simplifications to the overall serialization
algorithm to better reflect the 2.1 object model and simplify the 
implementation to make it easier to understand and maintain.
  (a) We now always send just the serialization id 0 when sending a null.
  (b) All subclasses of x10:;lang::Reference use a uniform serialization 
      scheme that requires the sending of a serialization id 
      (independent of whether or not the static type is a final class).
      Eliminates quite a bit of templatized code.
  (c) As a result of (a) and (b) and changes in the 2.1 object model, 
      we no longer need to send any bytes of address when serializing a ref.

Overall effect of these changes is to reduce the number of bytes sent for
an instance of x10::lang::Reference by 4 bytes on a 64 bit system; no
change in total bytes on 32 bit system.

Serialization for structs (whose static type is a struct) remains unchanged.
No serialization id and no extra bytes beyond the actual fields is sent.

If we do want to reduce bytes per class instance, I think the most productive
approach is to go for variable length serialization ids instead of attempting
to avoid sending the serialization id entirely.