Fix test case to match modern semantics.
Block dist is distributed in a block fashion as evenly as possible 
along a selected axis, not in a blocked fashion as evenly as possible 
among the total number of points in the Region. The previous test conditons
resulted in expecting some Regions to be blocked in the middle of a 
single value for an axis, which is not what the current specification
of Block dist does.