Fix 2275.
Ensure that when called on a ConstrainedType ct, xclause and realclause functions always return a CConstraint whose selfVar is the same as ct's selfvar. This ensures that methods on X10TypeMixin that look to find properties on a type, such as self.region, self.rank etc, access the correct "self". Subsequently when constraints such as self.rank==2 are added, the self refers to the correct self, and not the self variable of the returned xclause or realClause.

Also upgraded methods such as the ones that find properties to require a ConstrainedType so that any self terms that are returned will refer to the correct self. 

Moved some of the methods in X10Binary_c that "know" about certain kinds of operators to a separate file, BuiltinTypeRules. The goal is to get rid of this file -- that is to not have any special constraint propagation rules built into the type checker. But this requires support in the constraint system for conditional constraints, or constraints of the form x == (y == z), e.g.

class IntRange {

   def this(i:Int, j:Int):IntRange{self.zeroBased == (i == 0)} { ...}
   ....
   }
   
   Added comments to various methods in CConstraint, documenting which methods update the current constraint, and that CConstraint.copy() returns a constraint with the same selfVar as the original constraint.