Phase III
(1) Merged ObjectType and ReferenceType. Need a better name for the new type (is ObjectType for now).
(2) Renamed ArrayType to JavaArrayType. (Got rid of intfc/single implementing class distinction)
(3) Got rid of X10NamedType
(4) Got rid of ParameterType_c, moved it to ParameterType
(5) Got rid of ConstrainedType_c, moved it to ConstrainedType.
(6) Got rid of PrimitiveType_c, moved it to PrimitiveType.